pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-parser"
version = "0.1.0"
description = "Modular document parsing toolkit for Retrieval-Augmented Generation"
readme = "README.md"
authors = [{name = "RAG Parser Team"}]
requires-python = ">=3.9"
dependencies = ["pytest>=7.0"]

[project.optional-dependencies]
pdf = ["PyMuPDF>=1.22.0", "pdfplumber>=0.10.0"]
docs = ["python-docx>=0.8.11", "docx2python>=2.5.0"]
spreadsheets = ["pandas>=1.5.0", "openpyxl>=3.1.0"]
pptx = ["python-pptx>=0.6.21"]
ocr = ["paddleocr>=2.6.1.3", "pytesseract>=0.3.10", "Pillow>=9.0.0"]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
addopts = "-ra"
